Refactor Kernel/app into more encapsulated layout
